+++ sdmx-dl-format-base/src/main/java/internal/sdmxdl/format/NoOpPersistence.java
@@ -1,36 +1,32 @@
-package sdmxdl.format.kryo;
+package internal.sdmxdl.format;
 
 import lombok.NonNull;
-import nbbrd.service.ServiceProvider;
 import sdmxdl.DataRepository;
 import sdmxdl.format.FileFormat;
 import sdmxdl.format.spi.Persistence;
 import sdmxdl.web.MonitorReports;
 
-@ServiceProvider(Persistence.class)
-public final class KryoProvider implements Persistence {
+public enum NoOpPersistence implements Persistence {
 
-    private static final String ID = "KRYO";
-
-    private static final int RANK = 400;
+    INSTANCE;
 
     @Override
     public @NonNull String getPersistenceId() {
-        return ID;
+        return "NO_OP";
     }
 
     @Override
     public int getPersistenceRank() {
-        return RANK;
+        return UNKNOWN_PERSISTENCE_RANK;
     }
 
     @Override
     public @NonNull FileFormat<MonitorReports> getMonitorReportsFormat() throws IllegalArgumentException {
-        return FileFormat.of(new KryoFileFormat<>(MonitorReports.class), ".kryo");
+        return FileFormat.noOp();
     }
 
     @Override
     public @NonNull FileFormat<DataRepository> getDataRepositoryFormat() throws IllegalArgumentException {
-        return FileFormat.of(new KryoFileFormat<>(DataRepository.class), ".kryo");
+        return FileFormat.noOp();
     }
 }