+++ sdmx-dl-provider-base/src/main/java/sdmxdl/provider/PropertiesSupport.java
@@ -2,8 +2,10 @@ package sdmxdl.provider;
 
 import lombok.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
+import sdmxdl.file.SdmxFileSource;
 import sdmxdl.web.SdmxWebSource;
 
+import java.util.Collections;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Properties;
@@ -11,6 +13,10 @@ import java.util.function.Function;
 
 @lombok.experimental.UtilityClass
 public class PropertiesSupport {
+
+    public static @NonNull Function<? super String, ? extends CharSequence> asFunction(@NonNull SdmxFileSource source) {
+        return key -> PropertiesSupport.getProperty(Collections.emptyMap(), key);
+    }
 
     public static @NonNull Function<? super String, ? extends CharSequence> asFunction(@NonNull SdmxWebSource source) {
         return key -> PropertiesSupport.getProperty(source.getProperties(), key);