+++ CHANGELOG.md
@@ -9,8 +9,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- ![PROVIDER] Allow networking configuration by environment variables
-- ![PROVIDER] Allow networking configuration per source [#493](https://github.com/nbbrd/sdmx-dl/issues/493)
+- ![PROVIDER] Allow caching & networking configuration by environment variables [#516](https://github.com/nbbrd/sdmx-dl/issues/516)
+- ![PROVIDER] Allow caching & networking configuration per source [#493](https://github.com/nbbrd/sdmx-dl/issues/493)
 
 ### Fixed